crewai-x-agentops/example_1/pyproject.toml
[project]
name = "example-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "agentops>=0.3.21",
    "crewai>=0.86.0",
    "pymarkdownlnt>=0.9.26",
    "python-dotenv>=1.0.1",
]
